Implement host integration boundary with host functions and registry
- Added `host.rs` to define the integration boundary for native functionality
  exposed to the Mech runtime, including filesystem, database, UI, and network APIs.
- Introduced the `HostFunction` trait for callable host functions and a
  `ClosureHostFunction` struct for simple closure-backed implementations.
- Created `HostRegistry` trait and `InMemoryHostRegistry` for managing host
  functions, including registration, retrieval, and listing.
- Implemented `HostCallPolicy` trait for validating host function calls.
- Added utility functions for host calls and default resource/operation keys.
- Included error handling for invalid host functions and calls.
- Updated tests to cover registry functionality and host call validation.

Additionally, removed unused `host_arg` module from the runtime library.
